+++ front-end/nycroute-app/src/stationData.js
@@ -1,0 +1,101 @@
+// temp mock data
+const stationData = [
+  {
+    name: "8 St - NYU",
+    lines: ["R", "W"],
+  },
+  {
+    name: "West 4 St - Washington Square",
+    lines: ["A", "C"],
+  },
+  {
+    name: "14 St - Union Square",
+    lines: ["R", "W", "L", "6"],
+  },
+  {
+    name: "Astor Pl",
+    lines: ["4", "5", "6"],
+  },
+  {
+    name: "Times Square - 42 St",
+    lines: ["A", "C", "E", "N"], // a lot more in reality
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+  {
+    name: "Lorem Ispum Dolor Sit Amet",
+    lines: ["A", "B", "C", "1", "2", "3"],
+  },
+];
+
+export default stationData;
